backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.5
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
